Move temporary test helper functions to utils,js
